\name{CareerTrajectoryBattingMany}
\alias{CareerTrajectoryBattingMany}

\title{
Shiny App for Plotting Career Trajectories for Many Contemporary Batters
}
\description{
  Shiny app to graph career trajectories for a group of batters who satisfy particular criteria and measure of interest.
}
\usage{
  CareerTrajectoryBattingMany()
}
\author{
  Jim Albert
}


