\name{CareerTrajectoryFieldingMany}
\alias{CareerTrajectoryFieldingMany}

\title{
Shiny App for Plotting Career Trajectories for Many Contemporary Fielders
}
\description{
  Shiny app to graph career trajectories for a group of fielders at a specific position who satisfy particular criteria and measure of interest.
}
\usage{
  CareerTrajectoryFieldingMany()
}
\author{
  Jim Albert
}


