\name{CareerTrajectoryPitchingMany}
\alias{CareerTrajectoryPitchingMany}

\title{
Shiny App for Plotting Career Trajectories for Many Contemporary Pitchers
}
\description{
  Shiny app to graph career trajectories for a group of pitchers who satisfy particular criteria and measure of interest.
}
\usage{
  CareerTrajectoryPitchingMany()
}
\author{
  Jim Albert
}


