\name{FanGraphsBatting}
\alias{FanGraphsBatting}

\title{
Shiny App for Plotting Batting Career Trajectories Using FanGraphs Data
}
\description{
  Shiny app to graph career trajectories for a group of batters who satisfy particular criteria and measure of interest using FanGraphs data.
}
\usage{
  FanGraphsBatting()
}
\author{
  Jim Albert
}


