\name{FanGraphsPitching}
\alias{FanGraphsPitching}

\title{
Shiny App for Plotting Pitching Career Trajectories Using FanGraphs Data
}
\description{
  Shiny app to graph career trajectories for a group of pitchers who satisfy particular criteria and measure of interest using FanGraphs data.
}
\usage{
  FanGraphsPitching()
}
\author{
  Jim Albert
}


